split keyvaulAdapter from main.go
